Core: Clarify meaning of RSEProtocolPriorityError #6312
There is no implemented limit to how large a protocol priority can be.

Testing: Denote top-priority protocols with 0 instead of 1 #6312

Also shift down other enabled protocols.

Core: Make 0 the top priority #6312

Following #6201, both None and 0 denoted a disabled protocol. With this
commit, only None continues to do so, and any non-negative number
denotes an enabled protocol (with 0 being the highest priority).

This change may mean a different interpretation of the configuration of
existing Rucio instances, as well as affect backwards-compatibility with
previous versions of the Rucio clients.

Database: Remove RSE protocol default constraint #6312

When a new protocol is added to the database, if any of the priorities
are not defined, then set them to NULL. Previously, they were being set
to 0.

Core: Remove protocol priority entries from RSE settings #6312

Supposedly, these entries exist to specify which index refers to the top
priority protocol. However, it is unlikely that this was understood or
that any external system is using it.

Deletion: Rewrite function #6312

The original implementation contained two subtle assumptions: that there
is a single protocol for a particular scheme, and that there is always a
protocol with priority 1. These are common, but not necessarily true.

The new implementation sorts the protocols by their priority, then picks
the one with the highest priority.

Testing: Denote disabled protocols with None instead of 0 #6312

This was an omission from #6201.

Database: Rectify protocol priority mapping #6312

Specify that it may be None. This was an omission from #6201.